Added files handwriting_analysis.py and Readme.md
This bit of code trains data from the MNIST dataset from keras and stops when accuracy of 99% is reached. Note that in the run before upload, this code stopped after the 7th epoch because it reached an accuracy of 99.14%. For more information about this project, please check out the readme.